requirements.txt
# Core
langchain
langchain-openai
sentence-transformers
faiss-cpu
streamlit

# PDF parsing
pymupdf

# Optional LLM APIs
openai
python-dotenv
